modal/requirements.txt
modal==0.66.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
pydantic==2.9.2
google-genai==0.7.0
requests==2.32.3
pillow==10.4.0
